Agora Soma Double

+++ SomaSimples.java
@@ -5,12 +5,12 @@ public class SomaSimples {
     public static void main(String[] args) {
         Scanner ler = new Scanner(System.in);
         System.out.println("Digite Dois numeros:");
-        int a = 0;
-        int b = 0;
-        a = ler.nextInt();
-        b = ler.nextInt();
-        int soma = a + b;
+        double a = 0;
+        double b = 0;
+        a = ler.nextDouble();
+        b = ler.nextDouble();
+        double soma = a + b;
         System.out.println("SOMA = "+soma);
-
+        
     }
 }
